fix: 供电上限收敛到 int 语义（sensei 指出 Long.MAX 溢出疑虑）
Forge IEnergyStorage 是 int：机器缺口 ≤ int.MAX，cap 设再大也等效。
- feederPowerFeCap 范围 Long.MAX → Integer.MAX_VALUE
- 无上限(2速度卡) = Integer.MAX_VALUE（每轮灌满缺口，非 Long.MAX）
- ×16 用 long 域乘防溢出后钳到 int.MAX
- compat 默认 perPassCap Long.MAX → Integer.MAX_VALUE
- 界面条目范围/文案同步

+++ src/main/java/com/ae2addon/compat/AppFluxPowerCompat.java
@@ -68,17 +68,17 @@ public final class AppFluxPowerCompat {
     /** 给机器充能（多轮，默认每轮无上限）；兼容旧调用。 */
     public static long feedEnergy(BlockEntity target, Direction side,
             appeng.api.networking.IGrid grid, IActionSource source, int passes) {
-        return feedEnergy(target, side, grid, source, passes, Long.MAX_VALUE);
+        return feedEnergy(target, side, grid, source, passes, Integer.MAX_VALUE);
     }
 
     /** 给机器充能（单轮）；保留原签名兼容旧调用。 */
     public static long feedEnergy(BlockEntity target, Direction side,
             appeng.api.networking.IGrid grid, IActionSource source) {
-        return feedEnergy(target, side, grid, source, 1, Long.MAX_VALUE);
+        return feedEnergy(target, side, grid, source, 1, Integer.MAX_VALUE);
     }
 
     /**
-     * 单轮充能：上限 perPassCap（long；无上限传 Long.MAX）。
+     * 单轮充能：上限 perPassCap（long；≥ int.MAX 等效无上限）。
      * 机器能量槽是 int 容量，缺口本身 ≤ int max；cap 超过缺口时等价灌满缺口。
      */
     private static long feedEnergyOnce(BlockEntity target, Direction side,
